discorddumbstuff/pyproject.toml
[tool.poetry]
name = "discorddumbstuff"
version = "0.1.0"
description = "You know"
authors = ["megahomyak <g.megahomyak@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "1.7.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
